Simplify namespaces
