+++ src/Geode.Benchmarks/Program.cs
@@ -1,7 +1,5 @@
 ﻿global using BenchmarkDotNet.Running;
 global using BenchmarkDotNet.Attributes;
-global using Geode.Geometry;
-global using Geode.Algorithms;
-global using Geode.Serializers;
+global using Geode;
 
 BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
