+++ src/Geode.Tests/FeatureTests/CreateFeatureAsIFeatureConvertibleTests.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using System.Collections.Generic;
 using Xunit;
-using Geode.Geometry;
+using Geode;
 using Geode.Tests.FeatureTests.Models;
