+++ src/Geode.Tests/FeatureTests/Models/Geometries/Location.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using Geode.Geometry;
 
 namespace Geode.Tests.FeatureTests.Models.Geometries
 {
