+++ src/Geode/Geometry/BoundsExtensions.cs
@@ -2,9 +2,8 @@
 using System.Collections.Generic;
 using System.Text;
 using System.Linq;
-using Geode.Geometry;
 
-namespace Geode.Geometry
+namespace Geode
 {
     public static class BoundsExtensions
     {
