+++ src/Geode/Geometry/Feature.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using Geode.Geometry;
 using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Geode.Tests")]
