+++ src/Geode/Geometry/FeatureCollection.cs
@@ -3,7 +3,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using Geode.Geometry;
 
 namespace Geode
 {
