+++ src/Geode/Geometry/GeoJsonFeature.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Linq;
 
-namespace Geode.Geometry
+namespace Geode
 {
     public class GeoJsonFeature
     {
