+++ src/Geode/Geometry/GeometryCollection.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Geode.Geometry
+namespace Geode
 {
     public class GeometryCollection: IGeoType, IGeoCollection, IEnumerable<IGeoType>
     {
