+++ src/Geode/Geometry/Interfaces/IGeoType.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Collections;
 using System.Text;
 
-namespace Geode.Geometry
+namespace Geode
 {
     public interface IGeoType
     {
