+++ src/Geode/Geometry/MultiPolyline.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Collections;
 using System.Text;
 
-namespace Geode.Geometry
+namespace Geode
 {
     public class MultiPolyline : IGeoType, IGeometry
     {
