+++ src/Geode/Geometry/Point.cs
@@ -4,7 +4,7 @@ using System.Collections;
 using System.Text;
 using System.Linq;
 
-namespace Geode.Geometry
+namespace Geode
 {
     public sealed class Point : IGeoType, IPosition, IGeometry, IEnumerable<double>, IEquatable<Point>
     {
