+++ src/Geode/Geometry/PolylineExtensions.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Linq;
 
-namespace Geode.Geometry
+namespace Geode
 {
     public static class PolylineExtensions
     {
